Move standalone MFA enrollment start to the client API (#278)
The IS_USER standalone-enrollment entry point moves out of the flow API into
the client API: new ClientMfaEnrollmentController at
POST /api/v1/client/mfa/enrollment. MfaEnrollmentController (flow API) keeps
only the state-secured enrollment selection (GET) and skip (GET /skip) steps,
so it is no longer a mix of IS_USER and HAS_STATE security.
