+++ server/src/main/kotlin/com/sympauthy/api/controller/flow/MfaEnrollmentController.kt
@@ -1,27 +1,16 @@
 package com.sympauthy.api.controller.flow
 
 import com.sympauthy.api.controller.flow.auth.InteractiveAuthFlowSessionControllerUtil
-import com.sympauthy.api.resource.flow.MfaEnrollmentFlowResource
-import com.sympauthy.api.resource.flow.MfaEnrollmentInputResource
 import com.sympauthy.api.resource.flow.MfaFlowResource
 import com.sympauthy.api.resource.flow.MfaMethodResource
-import com.sympauthy.business.manager.ClientManager
-import com.sympauthy.business.manager.flow.InteractiveFlowEngine
-import com.sympauthy.business.manager.flow.InteractiveFlowSessionManager
-import com.sympauthy.business.manager.flow.auth.InteractiveAuthFlowSessionManager
 import com.sympauthy.business.manager.flow.mfa.InteractiveFlowSessionMfaEnrollmentManager
 import com.sympauthy.business.manager.flow.mfa.MfaAutoRedirect
 import com.sympauthy.business.manager.flow.mfa.MfaMethodSelection
 import com.sympauthy.security.SecurityRule.HAS_STATE
-import com.sympauthy.security.SecurityRule.IS_USER
-import com.sympauthy.security.clientId
 import com.sympauthy.security.stateOrNull
-import com.sympauthy.security.userId
 import io.micronaut.http.HttpResponse
-import io.micronaut.http.annotation.Body
 import io.micronaut.http.annotation.Controller
 import io.micronaut.http.annotation.Get
-import io.micronaut.http.annotation.Post
 import io.micronaut.security.annotation.Secured
 import io.micronaut.security.authentication.Authentication
 import io.swagger.v3.oas.annotations.Operation
@@ -29,66 +18,21 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import jakarta.inject.Inject
 
 /**
- * Controller owning MFA enrollment entry points:
- * - starting a standalone enrollment for a signed-in end-user, outside of an OAuth2 authorization (secured by
- *   the end-user's access token).
- * - skipping an optional enrollment offered during a flow (secured by the session state).
+ * Controller owning the MFA enrollment steps of an interactive flow (secured by the session state):
+ * - the enrollment method-selection step (choose the method to enroll, and skip when optional).
+ * - skipping an optional enrollment offered during a flow.
  *
- * Starting an enrollment creates an [com.sympauthy.business.model.flow.InteractiveFlowPurpose.MFA_ENROLLMENT]
- * session for the authenticated user and hands the UI the initial `state` + the enrollment page URL; once the
- * enrollment completes, the end-user is redirected back to the caller-provided `return_uri`.
+ * A standalone, client-initiated enrollment is started through the client API
+ * ([com.sympauthy.api.controller.client.ClientMfaEnrollmentController]); its subsequent steps are then driven
+ * here.
  */
-@Secured(IS_USER)
+@Secured(HAS_STATE)
 @Controller(MfaEnrollmentController.MFA_ENROLLMENT_ENDPOINT)
 class MfaEnrollmentController(
-    @Inject private val interactiveAuthFlowSessionManager: InteractiveAuthFlowSessionManager,
-    @Inject private val clientManager: ClientManager,
     @Inject private val mfaEnrollmentManager: InteractiveFlowSessionMfaEnrollmentManager,
-    @Inject private val engine: InteractiveFlowEngine,
     @Inject private val stepUriMapper: InteractiveFlowStepUriMapper,
-    @Inject private val sessionManager: InteractiveFlowSessionManager,
     @Inject private val interactiveAuthFlowSessionControllerUtil: InteractiveAuthFlowSessionControllerUtil,
 ) {
-
-    @Operation(
-        description = """
-Starts a standalone MFA enrollment for the authenticated end-user.
-
-Validates `return_uri` against the calling client's registered redirect URIs, creates an enrollment session,
-and returns the signed `state` and the `redirect_url` the UI must navigate the browser to. Once enrollment
-completes, the end-user is redirected to `return_uri`.
-        """,
-        responses = [
-            ApiResponse(
-                responseCode = "200",
-                description = "The enrollment session was started.",
-                useReturnTypeSchema = true
-            )
-        ]
-    )
-    @Post
-    suspend fun startEnrollment(
-        authentication: Authentication,
-        @Body resource: MfaEnrollmentInputResource
-    ): MfaEnrollmentFlowResource {
-        // Validate the return URI against the token's client registered redirect URIs to avoid open redirects.
-        val client = clientManager.findClientById(authentication.clientId)
-        val returnUri = interactiveAuthFlowSessionManager.parseRequestedRedirectUri(client, resource.returnUri)
-
-        val flow = interactiveAuthFlowSessionManager.getDefaultInteractiveFlow()
-        val session = mfaEnrollmentManager.startMfaEnrollmentSession(
-            userId = authentication.userId,
-            returnUri = returnUri,
-            flow = flow
-        )
-
-        val (steppedSession, step) = engine.getCurrentStep(session)
-        val redirectUri = stepUriMapper.toRedirectUri(steppedSession, flow, step)
-        return MfaEnrollmentFlowResource(
-            state = sessionManager.encodeState(steppedSession),
-            redirectUrl = redirectUri.toString()
-        )
-    }
 
     @Operation(
         description = """
@@ -111,7 +55,6 @@ Returns one of two response shapes:
         ],
         tags = ["flow"]
     )
-    @Secured(HAS_STATE)
     @Get
     suspend fun getEnrollmentSelection(
         authentication: Authentication
@@ -153,7 +96,6 @@ client-initiated one. The MFA step is marked as resolved so the flow does not pr
         ],
         tags = ["flow"]
     )
-    @Secured(HAS_STATE)
     @Get("/skip")
     suspend fun skipEnrollment(
         authentication: Authentication
